requirements.txt
azure-ai
tkinter
```